requirements.txt
streamlit>=1.36.0
google-generativeai>=0.7.2
faiss-cpu>=1.8.0.post1
pymupdf>=1.24.7
python-dotenv>=1.0.1
